+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/_document_model_administration_client.py
@@ -25,7 +25,7 @@ from ._polling import (
 from ._form_base_client import FormRecognizerClientBase
 from ._document_analysis_client import DocumentAnalysisClient
 from ._models import (
-    DocumentBuildMode,
+    ModelBuildMode,
     DocumentModelDetails,
     DocumentModelSummary,
     DocumentModelOperationDetails,
@@ -93,7 +93,7 @@ class DocumentModelAdministrationClient(FormRecognizerClientBase):
 
     @distributed_trace
     def begin_build_model(
-        self, build_mode: Union[str, DocumentBuildMode], blob_container_url: str, **kwargs: Any
+        self, build_mode: Union[str, ModelBuildMode], blob_container_url: str, **kwargs: Any
     ) -> DocumentModelAdministrationLROPoller[DocumentModelDetails]:
         """Build a custom model.
 
@@ -110,7 +110,7 @@ class DocumentModelAdministrationClient(FormRecognizerClientBase):
         :param str blob_container_url: An Azure Storage blob container's SAS URI. A container URI (without SAS)
             can be used if the container is public or has a managed identity configured. For more information on
             setting up a training data set, see: https://aka.ms/azsdk/formrecognizer/buildtrainingset.
-        :type build_mode: str or :class:`~azure.ai.formrecognizer.DocumentBuildMode`
+        :type build_mode: str or :class:`~azure.ai.formrecognizer.ModelBuildMode`
         :keyword str model_id: A unique ID for your model. If not specified, a model ID will be created for you.
         :keyword str description: An optional description to add to the model.
         :keyword str prefix: A case-sensitive prefix string to filter documents in the blob container url path.